@@ -2,18 +2,17 @@ using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 using System.Runtime.InteropServices.JavaScript;
 using System.Runtime.Versioning;
 
-namespace DualDrill.Client
+namespace DualDrill.Client;
+
+[SupportedOSPlatform("browser")]
+internal class Program
 {
-    [SupportedOSPlatform("browser")]
-    internal class Program
+    static async Task Main(string[] args)
     {
-        static async Task Main(string[] args)
-        {
-            var builder = WebAssemblyHostBuilder.CreateDefault(args);
-            builder.Services.AddSingleton<SignalRService>();
-            var app = builder.Build();
-            using var clientModule = await JSHost.ImportAsync("client", "/client.js");
-            await app.RunAsync();
-        }
+        var builder = WebAssemblyHostBuilder.CreateDefault(args);
+        builder.Services.AddSingleton<DualDrillBrowserSignalRClientService>();
+        var app = builder.Build();
+        using var clientModule = await JSHost.ImportAsync("client", "/client.js");
+        await app.RunAsync();
     }
 }
