@@ -7,7 +7,7 @@ namespace DualDrill.Client;
 public class SignalRService(NavigationManager NavigationManager) : IAsyncDisposable
 {
     readonly HubConnection Connection = new HubConnectionBuilder()
-        .WithUrl($"{NavigationManager.BaseUri}hub/user-input")
+        .WithUrl($"{NavigationManager.BaseUri}hub/browser-client")
         .Build();
 
     public async ValueTask DisposeAsync()
