@@ -1,5 +1,4 @@
-﻿@page "/desktop"
-@using Shared;
+﻿@using Shared;
 @rendermode @(new InteractiveServerRenderMode(prerender: false))
 
 <h3>Desktop Client</h3>
@@ -7,6 +6,7 @@
 {
     <span>Connecting ... </span>
 }
+<span>InteropMessage: @InteropMessage</span>
 <button @onclick="StartRTC">Start RTC</button>
 <div>
     <span> scale </span>
