@@ -2,4 +2,4 @@
 <h3>Blazor mix mode test</h3>
 
 <ServerSideCounter></ServerSideCounter>
-<DualDrill.Client.Pages.Counter></DualDrill.Client.Pages.Counter>
+<DualDrill.Client.Pages.ClientMain></DualDrill.Client.Pages.ClientMain>
